requirements.txt
# Backend API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.8.0

# LLM/STT integrations
requests>=2.32.0
openai-whisper>=20240930
torch>=2.2.0

# Audio and signal processing
sounddevice>=0.4.7
scipy>=1.12.0
numpy>=1.26.0
